pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kahunas-client"
version = "0.1.0"
description = "Python client, CLI, and MCP server for the Kahunas fitness API"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "httpx>=0.27,<1.0",
    "pydantic>=2.6,<3.0",
    "pydantic-settings>=2.2,<3.0",
    "pyyaml>=6.0",
    "tenacity>=8.2,<10.0",
    "fastmcp>=3.1,<4.0",
    "click>=8.1,<9.0",
    "rich>=13.0,<14.0",
    "openpyxl>=3.1,<4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "respx>=0.21",
    "mypy>=1.8",
    "ruff>=0.3",
    "types-PyYAML>=6.0",
]

[project.scripts]
kahunas = "kahunas_client.cli.main:cli"
kahunas-mcp = "kahunas_client.mcp.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/kahunas_client"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific
]
ignore = ["UP046"]  # Generic[T] subclass is needed for Pydantic models

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["TCH002", "TCH003"]  # don't require TYPE_CHECKING blocks in tests

[tool.ruff.lint.isort]
known-first-party = ["kahunas_client"]
